requirements.txt
Flask_SQLAlchemy
lxml~=5.4.0
matplotlib~=3.9.4
PyMySQL~=1.1.1
wordcloud~=1.9.4
SQLAlchemy~=2.0.41
Flask~=3.1.1
tqdm~=4.67.1
requests~=2.32.3
numpy~=1.26.4
apscheduler~=3.11.0
beautifulsoup4~=4.13.4
flask_wtf
jieba~=0.42.1
Pillow~=10.4.0
snownlp~=0.12.3
WTForms~=3.2.1
jinja2
werkzeug
itsdangerous
mysql-connector-python

scikit-learn~=1.2.2
torch~=2.1.2
shortuuid~=1.0.13
transformers~=4.51.3
llava~=1.2.2.post1
pandas~=2.2.3
uvicorn~=0.34.2
fastapi~=0.115.12
gradio~=4.16.0
tokenizers~=0.21.1
packaging~=25.0
datasets~=3.6.0
peft~=0.15.2
selenium~=4.32.0
webdriver-manager~=4.0.2
huggingface-hub~=0.31.4